+++ examples/global_variable/main.tf
@@ -9,22 +9,18 @@ terraform {
 
 provider "bamboo" {}
 
-//data "bamboo_global_variables" "variables" {}
+data "bamboo_global_variables" "variables" {}
 
-resource "bamboo_global_variable" "repo-user" {
-  key = "RepoUser2"
+resource "bamboo_global_variable" "nexus_user" {
+  key = "NexusUser"
   value = "nexus-user"
 }
 
 resource "bamboo_global_variable" "database" {
-  key = "Database4"
-  value = "Connection4"
+  key = "Database"
+  value = "Server=myServerAddress;Database=myDataBase;Uid=myUsername;Pwd=myPassword;"
 }
 
-//output "variables" {
-//  value = data.bamboo_global_variables.variables.global_variables
-//}
-
-# output "repo-user" {
-#   value = bamboo_global_variable.repo-user
-# }
+output "variables" {
+  value = data.bamboo_global_variables.variables.global_variables
+}