+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 FEATURES:
 
-- New Resource: ```bamboo_global_secret``` resource
+- New Resource: ```bamboo_global_variable``` resource